hangman-web-py/requirements.txt
flask==1.1.2
pyzmq==19.0.0
gevent==1.5.0
sawtooth-sdk==1.2.3
Flask-Sockets==0.2.1